test: split and refactor ProposalDetailCard.spec
