+++ frontend/svelte/src/tests/lib/components/proposal-detail/ProposalDetailCard/ProposalMeta.spec.ts
@@ -1,0 +1,54 @@
+/**
+ * @jest-environment jsdom
+ */
+
+import { Topic } from "@dfinity/nns";
+import { render } from "@testing-library/svelte";
+import ProposalMeta from "../../../../../lib/components/proposal-detail/ProposalDetailCard/ProposalMeta.svelte";
+import { mockProposalInfo } from "../../../../mocks/proposal.mock";
+
+const en = require("../../../../../lib/i18n/en.json");
+
+describe("ProposalMeta", () => {
+  it("should render proposal url", () => {
+    const { getByText } = render(ProposalMeta, {
+      props: {
+        proposalInfo: mockProposalInfo,
+      },
+    });
+    expect(getByText("url").getAttribute("href")).toBe("url");
+  });
+
+  it("should render proposer id", () => {
+    const { getByText } = render(ProposalMeta, {
+      props: {
+        proposalInfo: mockProposalInfo,
+      },
+    });
+    expect(
+      getByText(new RegExp(`${mockProposalInfo.proposer.toString()}$`))
+    ).toBeInTheDocument();
+  });
+
+  it("should render topic", () => {
+    const { getByText } = render(ProposalMeta, {
+      props: {
+        proposalInfo: mockProposalInfo,
+      },
+    });
+    expect(
+      getByText(new RegExp(`${en.topics[Topic[mockProposalInfo.topic]]}$`))
+    ).toBeInTheDocument();
+  });
+
+  it("should render id", () => {
+    const { getByText } = render(ProposalMeta, {
+      props: {
+        proposalInfo: mockProposalInfo,
+      },
+    });
+    expect(
+      getByText(new RegExp(`${mockProposalInfo.id.toString()}$`))
+    ).toBeInTheDocument();
+  });
+});
